QF3: improved  Qt 5.x compatibility
git-svn-id: https://www.dkfz.de/svn/B040/FCSTOOLS/trunk/QuickFit3@3746 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ lib/datatools.cpp
@@ -173,7 +173,7 @@ void dataExpand(QList<QList<QVariant> >& data, QStringList* columnsNames) {
                         else if (i==3) vo=v.value<QVector4D>().w();
                         break;
                     case QVariant::QVariant::Matrix4x4: {
-                            qreal* dd=v.value<QMatrix4x4>().data();
+                            float* dd=v.value<QMatrix4x4>().data();
                             if (i<16 && i>=0) vo=dd[i];
                         } break;
                     case QVariant::List: {
@@ -256,7 +256,7 @@ void dataReduce(QList<QList<QVariant> >& data, QStringList* columnsNames) {
                         else if (i==3) vo=v.value<QVector4D>().w();
                         break;
                     case QVariant::QVariant::Matrix4x4: {
-                            qreal* dd=v.value<QMatrix4x4>().data();
+                            float* dd=v.value<QMatrix4x4>().data();
                             if (i<16 && i>=0) vo=dd[i];
                         } break;
                     case QVariant::List: {
